+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `benf\neo\models\BlockType::getEntryType()`
 - Added `benf\neo\models\BlockType::setEntryType()`
 - Added `benf\neo\models\BlockType::entryTypeId`
+- Added `benf\neo\services\BlockTypes::getByCriteria()`
 - Added `benf\neo\services\Conversion::convertEntryTypeToBlockType()`
 
 ### Changed